udpate SDK version Update to SDK v1.16, change from framework to xcframework
