--- FlyBuySDK.framework/Modules/module.modulemap
@@ -1,11 +1,0 @@
-framework module FlyBuySDK {
-  umbrella header "FlyBuySDK.h"
-
-  export *
-  module * { export * }
-}
-
-module FlyBuySDK.Swift {
-    header "FlyBuySDK-Swift.h"
-    requires objc
-}
