+++ FlyBuySDK.xcframework/ios-x86_64-simulator/FlyBuySDK.framework/Modules/module.modulemap
@@ -1,0 +1,11 @@
+framework module FlyBuySDK {
+  umbrella header "FlyBuySDK.h"
+
+  export *
+  module * { export * }
+}
+
+module FlyBuySDK.Swift {
+    header "FlyBuySDK-Swift.h"
+    requires objc
+}
