+++ FlyBuySDK.xcframework/ios-x86_64-simulator/FlyBuySDK.framework/_CodeSignature/CodeResources
@@ -1,0 +1,223 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/FlyBuySDK-Swift.h</key>
+		<data>
+		+nH8K+qUuVVLPkTQNQt2PT5hbEg=
+		</data>
+		<key>Headers/FlyBuySDK.h</key>
+		<data>
+		1ym/faX6wAx0a1Qi4713yB/KwBg=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		6KZx3A2aZFwsg2s+vqdT2soTLzs=
+		</data>
+		<key>IrisModel.momd/FlyBuyModel.mom</key>
+		<data>
+		4OT59v6i2lTgTcWI63x5ejzsBNo=
+		</data>
+		<key>IrisModel.momd/VersionInfo.plist</key>
+		<data>
+		5VVbXQkOLykWcc4KdXDWm0K709E=
+		</data>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<data>
+		icfK5omZrUcJbRVjta2QLrt5jI0=
+		</data>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<data>
+		ku//H5QSZv5Jyet+rV6bSRUPGWk=
+		</data>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<data>
+		3LYpN3+EWXT4ArKWJd2vOqrPQiw=
+		</data>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64.swiftdoc</key>
+		<data>
+		icfK5omZrUcJbRVjta2QLrt5jI0=
+		</data>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64.swiftinterface</key>
+		<data>
+		ku//H5QSZv5Jyet+rV6bSRUPGWk=
+		</data>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64.swiftmodule</key>
+		<data>
+		3LYpN3+EWXT4ArKWJd2vOqrPQiw=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		gz6z+ujuZRxGF4uswrAmAccDd60=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/FlyBuySDK-Swift.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			039jRdz92ZMiPUxLNwetjdGFidqblJ0mLDrfr8h5yZc=
+			</data>
+		</dict>
+		<key>Headers/FlyBuySDK.h</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			t9TBcMliHbOYM9SRZ5HBLhTBsi4U8u2FlIQKhHXJwns=
+			</data>
+		</dict>
+		<key>IrisModel.momd/FlyBuyModel.mom</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			H6YWavwxKE/LVmdHHznCO6/V9jpp+gwhJzxdZJBEbb8=
+			</data>
+		</dict>
+		<key>IrisModel.momd/VersionInfo.plist</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			13eLnAj6niH0nT8Z2r3Fg97loiEwCfYsLvXX/lW7TK8=
+			</data>
+		</dict>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64-apple-ios-simulator.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			7uQJIG5ec8VSCdecJFTnBAOdtH40LIPQ6A9ab3WE/J4=
+			</data>
+		</dict>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64-apple-ios-simulator.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			dM8/t/Iy4ZMnCwa8M2tp7LscdVKpl2Iknz+E9EkUW8c=
+			</data>
+		</dict>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64-apple-ios-simulator.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			rl0IvwvyfXqGL/+CWPZPKgEy4U5D2evd9DuVl/6aacw=
+			</data>
+		</dict>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64.swiftdoc</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			7uQJIG5ec8VSCdecJFTnBAOdtH40LIPQ6A9ab3WE/J4=
+			</data>
+		</dict>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64.swiftinterface</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			dM8/t/Iy4ZMnCwa8M2tp7LscdVKpl2Iknz+E9EkUW8c=
+			</data>
+		</dict>
+		<key>Modules/FlyBuySDK.swiftmodule/x86_64.swiftmodule</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			rl0IvwvyfXqGL/+CWPZPKgEy4U5D2evd9DuVl/6aacw=
+			</data>
+		</dict>
+		<key>Modules/module.modulemap</key>
+		<dict>
+			<key>hash2</key>
+			<data>
+			eV+LQqm5tNrTEj1mUk7b8a4DjG3SkHlQfhyr6FRNXrM=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
